+++ src/components/Features/Features.module.scss
@@ -352,82 +352,6 @@ $mobile-horizontal-padding: 16px;
   h2 {
     @include features-caption();
   }
-
-  .frameworks_box {
-    margin-top: 60px;
-    width: 100%;
-    min-height: 656px;
-    background: #ffffff;
-    box-shadow: 0px 4px 16px rgba(16, 46, 52, 0.1);
-    border-radius: 16px;
-
-    @include breakpoint($desktop-size-breakpoint) {
-      width: 996px;
-    }
-
-    .frameworks_box_header {
-      display: flex;
-      justify-content: space-between;
-      height: 64px;
-      border-bottom: 1px solid #edeff7;
-      overflow: scroll;
-
-      div {
-        @include font-poppins($fw-semi-bold);
-        @include font-scale();
-
-        display: flex;
-        align-items: center;
-        justify-content: center;
-        height: 100%;
-        width: 100%;
-        margin: 0 20px 0 20px;
-        color: #8791ab;
-
-        &:hover {
-          cursor: pointer;
-        }
-      }
-
-      .active_lang {
-        border-bottom: 2px solid #009dbb;
-      }
-    }
-
-    .frameworks_box_content {
-      display: flex;
-      padding: 24px;
-      flex-flow: wrap;
-      justify-content: space-between;
-      gap: 15px;
-
-      &::after {
-        content: '';
-        width: 219px;
-      }
-
-      .frameworks_box_content_item {
-        margin-bottom: 24px;
-        padding: 0px 12px;
-        flex: 1;
-        height: 80px;
-        border: 1px solid #edeff7;
-        border-radius: 10px;
-        display: flex;
-        justify-content: center;
-
-        @include breakpoint($desktop-size-breakpoint) {
-          width: 219px;
-          flex: none;
-        }
-      }
-
-      img {
-        height: 100%;
-        width: 110px;
-      }
-    }
-  }
 }
 
 .subscription {
