+++ src/components/Features/dataSource.js
@@ -85,74 +85,5 @@ export const collapsableList = [
   },
 ];
 
-export const languageList = [
-  { lang: 'Java', id: 'java' },
-  { lang: '.NET', id: 'dotnet' },
-  { lang: 'JavaScript', id: 'javascript' },
-  { lang: 'Python', id: 'python' },
-  { lang: 'PHP', id: 'php' },
-  { lang: 'Other', id: 'other' },
-];
-
 export const customerIcons = [...Object.values(iconsCustomers)];
 export const frameworkIcons = [...Object.values(iconsFrameworks)];
-
-export const frameworkIconsJava = [
-  { icon: iconsFrameworks.testng },
-  { icon: iconsFrameworks.junit },
-  { icon: iconsFrameworks.junit5 },
-  { icon: iconsFrameworks.spock },
-  { icon: iconsFrameworks.soapui },
-  { icon: iconsFrameworks.jbehave },
-  { icon: iconsFrameworks.scalatest },
-  { icon: iconsFrameworks.cucumber },
-  { icon: iconsFrameworks.cucumber2 },
-  { icon: iconsFrameworks.cucumber3 },
-  { icon: iconsFrameworks.cucumber4 },
-  { icon: iconsFrameworks.cucumber5 },
-  { icon: iconsFrameworks.cucumber6 },
-  { icon: iconsFrameworks.serenity },
-  { icon: iconsFrameworks.karate },
-  { icon: iconsFrameworks.artos },
-  { icon: iconsFrameworks.android },
-];
-
-export const frameworkIconsDotNet = [
-  { icon: iconsFrameworks.nunit },
-  { icon: iconsFrameworks.vstest },
-  { icon: iconsFrameworks.specflow },
-  { icon: iconsFrameworks.unitnet },
-];
-
-export const frameworkIconsJavascript = [
-  { icon: iconsFrameworks.cucumber },
-  { icon: iconsFrameworks.jasmine },
-  { icon: iconsFrameworks.mocha },
-  { icon: iconsFrameworks.nightwatch },
-  { icon: iconsFrameworks.jest },
-  { icon: iconsFrameworks.cypress },
-  { icon: iconsFrameworks.codecept },
-  { icon: iconsFrameworks.postman },
-  { icon: iconsFrameworks.testcafe },
-  { icon: iconsFrameworks.webdriver },
-  { icon: iconsFrameworks.playwright },
-];
-
-export const frameworkIconsPython = [
-  { icon: iconsFrameworks.pytest },
-  { icon: iconsFrameworks.robotframework },
-  { icon: iconsFrameworks.pythonbehave },
-  { icon: iconsFrameworks.nose },
-];
-
-export const frameworkIconsPhp = [
-  { icon: iconsFrameworks.behat },
-  { icon: iconsFrameworks.phpunit },
-  { icon: iconsFrameworks.codeception },
-];
-
-export const frameworkIconsOther = [
-  { icon: iconsFrameworks.st },
-  { icon: iconsFrameworks.gwen },
-  { icon: iconsFrameworks.gauge },
-];
