+++ src/components/SupportedFrameworks/SupportedFrameworks.module.scss
@@ -1,0 +1,80 @@
+@import 'src/styles/mixins/font';
+@import 'src/styles/mixins/font-scale';
+@import 'src/styles/mixins/breakpoint';
+@import 'src/styles/variables';
+
+.frameworks_wrapper {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  text-align: center;
+
+  .frameworks_box {
+    margin-top: 60px;
+    width: 100%;
+    min-height: 656px;
+    background: var(--white);
+    box-shadow: 0px 4px 16px rgba(16, 46, 52, 0.1);
+    border-radius: 16px;
+
+    @include breakpoint($desktop-sm) {
+      width: 996px;
+    }
+
+    .frameworks_box_header {
+      display: flex;
+      justify-content: space-between;
+      height: 64px;
+      border-bottom: 1px solid var(--purple-grey);
+
+      div {
+        @include font-poppins($fw-semi-bold);
+        @include font-scale();
+
+        display: flex;
+        align-items: center;
+        justify-content: center;
+        height: 100%;
+        width: 100%;
+        margin: 0 20px 0 20px;
+        color: var(--text-service);
+
+        &:hover {
+          cursor: pointer;
+        }
+      }
+
+      .active_lang {
+        border-bottom: 2px solid var(--color-primary-600);
+      }
+    }
+
+    .frameworks_box_content {
+      display: grid;
+      grid-template-columns: repeat(2, 1fr);
+      gap: 15px;
+      padding: 24px;
+      
+      @include breakpoint($desktop-sm) {
+        grid-template-columns: repeat(4, 1fr);
+      }
+
+      .frameworks_box_content_item {
+        display: flex;
+        flex: 1 0 50px;
+        justify-content: center;
+        height: 80px;          
+        margin-bottom: 24px;
+        padding: 0px 12px;
+        border: 1px solid var(--purple-grey);
+        border-radius: 10px;
+      }
+
+      img {
+        height: 100%;
+        width: 100%;
+        max-width: 110px;
+      }
+    }
+  }
+}
